añadiendo eliminar imagen de producto en upload al eliminar el producto en la vista modificar producto

+++ controllers/admin/modificar.php
@@ -2,6 +2,7 @@
 include("curlController.php");
 include("logger.php");
 use Clases\CurlController;
+use PrestaShop\PrestaShop\Adapter\Entity\Module as EntityModule;
 
 class ModificarController extends ModuleAdminController{
 
@@ -12,6 +13,7 @@ class ModificarController extends ModuleAdminController{
         parent::__construct();
         $this->bootstrap = true;
         $this->id_idioma = $this->context->language->id;
+        $this->modulo= EntityModule::getInstanceByName("wbzalando");
     }
 
     public function init()
@@ -37,43 +39,6 @@ class ModificarController extends ModuleAdminController{
         $this->context->smarty->assign($variablesSmarty);
         $this->setTemplate('/modificar/vista.tpl');
     }
-
-    // function ajaxProcessGetEliminarProducto(){
-    //     $respuesta_servidor=["respuestaServidor" => []];
-    //     $eliminarPrecioDB=$this->eliminarPrecio($_GET["ean"],$_GET["idPais"]);
-    //     $eliminarStockDB=$this->eliminarStock($_GET["ean"],$_GET["idPais"]);
-    //     $consultarExistenciaProductoDB=$this->consultarPrecioPorEan($_GET["ean"]);
-    //     if(count($consultarExistenciaProductoDB)===0){
-    //         $respustaEliminarConfig=$this->eliminarConfig($_GET["idConfig"]);
-    //         $respuestaExistenciaConfigModelo=$this->consultarConfigPorModelo($_GET["idModelo"]);
-    //         if(count($respuestaExistenciaConfigModelo)===0){
-    //             $this->eliminarModelo($_GET["idModelo"]);
-    //         }
-    //         $respuesta_servidor["respuestaServidor"]=[
-    //             "datos" => $consultarExistenciaProductoDB,
-    //             "eliminarPrecioDB" => $eliminarPrecioDB,
-    //             "eliminarStockDB" => $eliminarStockDB,
-    //             "respustaEliminarConfig" => $respustaEliminarConfig
-    //         ];
-    //     }
-    //     // $respuestaDB=$this->eliminar($_GET["id"]);
-    //     // if($respuestaDB){
-    //     //     $respuesta_servidor["respuestaServidor"]=[
-    //     //         "mensaje" => "eliminacion cumpletada"
-    //     //     ];
-    //     // }
-    //     // else{
-    //     //     $respuesta_servidor["respuestaServidor"]=[
-    //     //         "mensaje" => "error al eliminar"
-    //     //     ];
-    //     // }
-    //     print(json_encode($respuesta_servidor));
-    // }
-
-    // // function eliminar($id){
-    // //     $SQL="DELETE FROM ps_wbzalando_modelo_producto WHERE id_modelo_producto ='".$id."';";
-    // //     return Db::getInstance()->execute($SQL);
-    // // }
 
     public function ajaxProcessGetConsultarPaisesZalando(){
         $respuesta_servidor=["respuestaServidor" => [],"estatuRespuestaApi" => 0];
@@ -406,9 +371,18 @@ class ModificarController extends ModuleAdminController{
     public function ajaxProcessPostEliminarProducto(){
         $respuesta_servidor=["respuestaServidor" => []];
         $respuestaEliminarProducto=$this->eliminarProducto($_POST["id_modelo"]);
-        foreach($_POST["eans"] as $ean){
-            $this->eliminarStock($ean["ean"],$ean["idPais"]);
-            $this->eliminarPrecio($ean["ean"],$ean["idPais"]);
+        if($respuestaEliminarProducto){
+            $nombreImagen=$_POST["nombre_imagen"];
+            $dir=_PS_MODULE_DIR_.$this->modulo->name."/upload/";
+            if(file_exists($dir.$nombreImagen)){
+                unlink($dir.$nombreImagen);
+            }
+        }
+        if($respuestaEliminarProducto){
+            foreach($_POST["eans"] as $ean){
+                $this->eliminarStock($ean["ean"],$ean["idPais"]);
+                $this->eliminarPrecio($ean["ean"],$ean["idPais"]);
+            }
         }
         $respuesta_servidor["respuestaServidor"]=[
             "producto_eliminada" => $respuestaEliminarProducto
